% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idb_applications.R
\name{idb_applications}
\alias{idb_applications}
\title{List applications}
\usage{
idb_applications(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a tibble
}
\description{
List applications
}
\examples{
\dontrun{
idb_applications()
}
}

