% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idb_indices.R
\name{idb_indices}
\alias{idb_indices}
\title{List indices}
\usage{
idb_indices(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a tibble
}
\description{
List indices
}
\examples{
\dontrun{
idb_indices()
}
}

