% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idb_search.R
\name{idb_search}
\alias{idb_search}
\title{Search remote sensing indices}
\usage{
idb_search(query, ...)
}
\arguments{
\item{query}{(character) query string}

\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
xxxx
}
\description{
Search remote sensing indices
}

