% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idb_sensors.R
\name{idb_sensors}
\alias{idb_sensors}
\title{List sensors}
\usage{
idb_sensors(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a tibble
}
\description{
List sensors
}
\examples{
\dontrun{
idb_sensors()
}
}

