% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridb-package.R
\docType{package}
\name{ridb-package}
\alias{ridb}
\alias{ridb-package}
\title{ridb - Index Database R client}
\description{
ridb - Index Database R client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

